Add files via upload
Its a little synthesizer to play around with a 31tet scale over a midipiano. If you want to play it you need a MIDI keyboard or similar device connected to your computer.The synth will listen to Midichannel "1" which should be standart. There are no Control-Change messages used. Consider this a work in progress =)